refactor: extract parameters.go & template.go from main.go
Not great encapsulation, but file-level isolation of imports is nice.
